+++ CloudShell.ResourceDefinitions.ResourceManager/ResourceModelResourceManagerObservabilityProvider.cs
@@ -1,0 +1,9 @@
+using CloudShell.Abstractions.ResourceManager;
+using ResourceModelResource = CloudShell.ResourceDefinitions.Resource;
+
+namespace CloudShell.ResourceDefinitions.ResourceManager;
+
+public interface IResourceModelResourceManagerObservabilityProvider
+{
+    ResourceObservability? GetObservability(ResourceModelResource resource);
+}